Address PR #86 feedback: move imports to top and simplify architecture
- Move IngestionPipeline import to top level in celery_tasks.py
- Move select, ImportRun, ImportStatus imports to top level in dedup.py
- Move EntityType-to-model mapping from registry.py to models/__init__.py
- Remove wrapper function from base.py that delegated to registry.py
- Delete redundant models/registry.py (models already imported in __init__)
- Update CLAUDE.md with coding guidelines about imports and architecture

No circular imports existed - the function-level imports were unnecessary.
